+++ plugins/DigitizerReco.h
@@ -20,7 +20,7 @@ public:
     ~DigitizerReco(){};
 
     //---utils---
-    bool Begin(CfgManager& opts, int* index);
+    bool Begin(CfgManager& opts, uint64* index);
     bool ProcessEvent(const H4Tree& event, map<string, PluginBase*>& plugins, CfgManager& opts);
     
 private:
