+++ plugins/HodoReco.h
@@ -21,7 +21,7 @@ public:
     ~HodoReco() {};
 
     //---utils---
-    bool Begin(CfgManager& opts, int* index);
+    bool Begin(CfgManager& opts, uint64* index);
     bool ProcessEvent(const H4Tree& event, map<string, PluginBase*>& plugins, CfgManager& opts);
     
 private:
