+++ plugins/MakeCovarianceMatrix.h
@@ -18,7 +18,7 @@ public:
     ~MakeCovarianceMatrix(){};
 
     //---utils---
-    bool Begin(CfgManager& opts, int* index);
+    bool Begin(CfgManager& opts, uint64* index);
     bool ProcessEvent(const H4Tree& event, map<string, PluginBase*>& plugins, CfgManager& opts);
     bool End(CfgManager& opts);
