+++ plugins/PluginBase.h
@@ -38,7 +38,7 @@ public:
     vector<SharedData> GetSharedData(string tag="", string type="", bool permanent=true);
     
     //---utils---
-    virtual bool Begin(CfgManager& opts, int* index){};
+    virtual bool Begin(CfgManager& opts, uint64* index){};
     virtual bool ProcessEvent(const H4Tree& event, map<string, PluginBase*>& plugins, CfgManager& opts){};
     virtual bool End(CfgManager& opts){};
