+++ plugins/WireChamberReco.h
@@ -16,7 +16,7 @@ public:
     ~WireReco() {};
 
     //---utils---
-    bool Begin(CfgManager& opts, int* index);
+    bool Begin(CfgManager& opts, uint64* index);
     bool ProcessEvent(const H4Tree& event, map<string, PluginBase*>& plugins, CfgManager& opts);
     
 private:
